libraryapi/requirement.txt
python-decouple
Django==1.11.4
django-filter==1.0.4
djangorestframework
djangorestframework-simplejwt=5.3.1
django-crontab
